+++ packages/editor-common/web/src/index.ts
@@ -2,26 +2,17 @@
 export * from './Icons';
 
 // Components
-export { default as InfoIcon } from './Components/InfoIcon';
-export { default as Checkbox } from './Components/Checkbox';
 export { default as ClickOutside } from './Components/ClickOutside/ClickOutside';
 export { default as useClickOutside } from './Components/ClickOutside/useClickOutside';
 
-export { default as Dropdown } from './Components/Dropdown';
-
-export { default as FocusManager } from './Components/FocusManager';
 export { default as LinkPanelWrapper } from './Components/LinkComponents/LinkPanelWrapper';
 export { default as LinkButton } from './Components/LinkComponents/LinkButton';
 export { default as LinkModal } from './Components/LinkComponents/LinkModal';
-export { default as RadioGroup } from './Components/RadioGroup';
-export { default as Separator } from './Components/Separator';
 export { default as ToolbarButton } from './Components/ToolbarButton';
 export { default as InlineToolbarButton } from './Components/InlineToolbarButton';
-export { default as TextSearchInput } from './Components/TextSearchInput';
 
 //Modals
 export { default as EditorModals } from './Modals/EditorModals';
-export { default as RichContentModal } from './Modals/RichContentModal';
 
 //Utils
 export { default as decorateComponentWithProps } from './Utils/decorateComponentWithProps';
@@ -81,8 +72,11 @@ export {
   isCursorAtStartOfContent,
   isCursorAtFirstLine,
   selectAllContent,
+  isAtomicBlockInSelection,
   setSelectionToBlock,
+  hasBlockType,
 } from './Utils/draftUtils';
+export { getColor, setTextColor, setHighlightColor } from './Utils/colorUtils';
 export { triggerMention, insertMention } from './Utils/mentionUtils';
 export { isiOS } from './Utils/isiOS';
 export { mergeToolbarSettings } from './Utils/mergeToolbarSettings';
@@ -92,7 +86,6 @@ export {
   MODIFIERS,
   TOOLBARS,
   DISPLAY_MODE,
-  DECORATION_MODE,
   CHARACTERS,
   FORMATTING_BUTTONS,
   INSERT_PLUGIN_BUTTONS,
@@ -154,3 +147,5 @@ export const emptyDraftContent = {
     },
   ],
 };
+
+export { getSelectionStyles } from './Utils/inlineStyleUtils';