+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -22,7 +22,6 @@ import {
   getBlockInfo,
   getFocusedBlockKey,
   createCalcContentDiff,
-  getEditorContentSummary,
   getBlockType,
   COMMANDS,
   MODIFIERS,
@@ -73,6 +72,7 @@ import {
   PluginEventParams,
   OnPluginAction,
   IMAGE_TYPE,
+  EditorCommands,
 } from 'wix-rich-content-common';
 import styles from '../../statics/styles/rich-content-editor.scss';
 import draftStyles from '../../statics/styles/draft.rtlignore.scss';
@@ -137,6 +137,7 @@ export interface RichContentEditorProps extends PartialDraftEditorProps {
   customAnchorScroll?: CustomAnchorScroll;
   style?: CSSProperties;
   locale: string;
+  localeContent?: string;
   shouldRenderOptimizedImages?: boolean;
   onChange?(editorState: EditorState): void;
   onAtomicBlockFocus?: onAtomicBlockFocus;
@@ -206,8 +207,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
 
   editorWrapper!: Element;
 
-  TiptapEditor: React.FC<{ editorProps; onUpdate }> | null;
-
   lastFocusedAtomicPlugin?: ContentBlock;
 
   updateBounds!: (editorBounds?: BoundingRect) => void;
@@ -224,7 +223,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
 
   innerRCECustomStyleFn;
 
-  EditorCommands!: ReturnType<typeof createEditorCommands>;
+  EditorCommands!: EditorCommands;
 
   getSelectedText!: (editorState: EditorState) => string;
 
@@ -238,15 +237,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     },
     normalize: {},
     plugins: [],
-  };
-
-  static publish = async (
-    postId: number,
-    editorState: EditorState,
-    callBack: (...args) => boolean = () => true
-  ) => {
-    const postSummary = getEditorContentSummary(editorState || {});
-    callBack({ postId, ...postSummary });
   };
 
   static getDerivedStateFromError(error: string) {
@@ -266,7 +256,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       undoRedoStackChanged: false,
     };
     this.refId = Math.floor(Math.random() * 9999);
-    this.TiptapEditor = null;
     this.commonPubsub = simplePubsub();
     this.handleCallbacks = this.createContentMutationEvents(
       this.state.editorState,
@@ -294,26 +283,12 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     this.handleBlockFocus(this.state.editorState);
   }
 
-  tiptapInterval;
-
   componentDidMount() {
     preventWixFocusRingAccessibility(this.editorWrapper);
     this.reportDebuggingInfo();
     this.preloadLibs();
     document?.addEventListener('beforeinput', this.preventDefaultKeyCommands);
     this.commonPubsub.set('undoExperiment', this.getUndoExperiment);
-
-    if (this.props.experiments?.tiptapEditor?.enabled) {
-      this.tiptapInterval = setTimeout(async () => {
-        const tiptapEditorModule = await import(
-          /* webpackChunkName: wix-tiptap-editor */
-          'wix-tiptap-editor'
-        );
-        const { TiptapEditor } = tiptapEditorModule;
-        this.TiptapEditor = TiptapEditor;
-        this.forceUpdate();
-      }, 1000);
-    }
   }
 
   componentWillMount() {
@@ -323,7 +298,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
   }
 
   componentWillUnmount() {
-    clearTimeout(this.tiptapInterval);
     this.updateBounds = () => '';
     document?.removeEventListener('beforeinput', this.preventDefaultKeyCommands);
   }
@@ -401,6 +375,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       theme,
       t,
       locale,
+      localeContent,
       anchorTarget,
       relValue,
       customAnchorScroll,
@@ -421,6 +396,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       theme: theme || {},
       t,
       locale,
+      localeContent,
       anchorTarget,
       relValue,
       customAnchorScroll,
@@ -475,6 +451,8 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
         onPluginAction,
         onPluginChange: (pluginId: string, changeObj) =>
           helpers.onPluginChange?.(pluginId, changeObj, Version.currentVersion),
+        onToolbarButtonClick: args =>
+          helpers.onToolbarButtonClick?.({ ...args, version: Version.currentVersion }),
       },
       config,
       isMobile,
@@ -531,6 +509,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     const { createPluginsDataMap = {} } = this.props;
     this.EditorCommands = createEditorCommands(
       createPluginsDataMap,
+      this.plugins,
       this.getEditorState,
       this.updateEditorState
     );
@@ -875,15 +854,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     pubsub: this.commonPubsub,
   });
 
-  // TODO: remove deprecated postId once getContent(postId) is removed (9.0.0)
-  publish = async (postId?: string) => {
-    if (!this.props.helpers?.onPublish) {
-      return;
-    }
-    const { pluginsCount, pluginsDetails } = getEditorContentSummary(this.state.editorState) || {};
-    this.props.helpers.onPublish(postId, pluginsCount, pluginsDetails, Version.currentVersion);
-  };
-
   setEditor = (ref: Editor) => (this.editor = get(ref, 'editor', ref));
 
   inPluginEditingMode = false;
@@ -977,15 +947,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     }
 
     return handled || 'not-handled';
-  };
-
-  renderTiptapEditor = () => {
-    if (this.TiptapEditor) {
-      const TiptapEditor = this.TiptapEditor;
-      return <TiptapEditor onUpdate={() => null} editorProps={this.props} />;
-    } else {
-      return null;
-    }
   };
 
   renderEditor = () => {
@@ -1201,7 +1162,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
         ...themeDesktopStyle,
       });
 
-      const isTiptapEditor = this.props.experiments?.tiptapEditor?.enabled;
       return (
         <GlobalContext.Provider value={this.state.context}>
           <Measure bounds onResize={this.onResize}>
@@ -1224,12 +1184,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
                 >
                   {this.renderAccessibilityListener()}
 
-                  {isTiptapEditor && this.renderTiptapEditor()}
-                  {isTiptapEditor && (
-                    <div style={{ position: 'absolute', zIndex: -1 }}> {this.renderEditor()}</div>
-                  )}
-
-                  {!isTiptapEditor && this.renderEditor()}
+                  {this.renderEditor()}
                   {showToolbars && this.renderToolbars()}
                   {this.renderInlineModals()}
                   {this.renderErrorToast()}