+++ packages/ricos-content/web/src/converters/tiptap/consts.ts
@@ -25,6 +25,8 @@ export const NODE_TYPES = [
   Node_Type.OEMBED,
 ];
 
+export const NODE_MAP = Object.fromEntries(NODE_TYPES.map(v => [v, v]));
+
 export const DECORATION_TYPES = [
   Decoration_Type.BOLD,
   Decoration_Type.ITALIC,