+++ packages/ricos-editor/web/src/utils/editorCommandsTestsUtil.ts
@@ -10,6 +10,8 @@ import {
   RICOS_IMAGE_TYPE,
   RICOS_LINK_TYPE,
   RICOS_MENTION_TYPE,
+  RICOS_TEXT_COLOR_TYPE,
+  RICOS_TEXT_HIGHLIGHT_TYPE,
 } from 'ricos-content';
 import {
   DividerData,
@@ -23,7 +25,6 @@ import {
   ImageData,
   LinkData,
   Node_Type,
-  Decoration_Type,
   MentionData,
   PluginContainerData_Width_Type,
   PluginContainerData_Alignment,
@@ -275,6 +276,7 @@ const gallery = {
       },
     ],
     styles: {
+      allowContextMenu: true,
       allowDownload: false,
       allowHover: true,
       allowSocial: false,
@@ -323,6 +325,7 @@ const gallery = {
       },
     ],
     styles: {
+      allowContextMenu: true,
       allowDownload: false,
       allowHover: true,
       allowSocial: false,
@@ -484,7 +487,6 @@ const image = {
 
 const link = {
   type: RICOS_LINK_TYPE,
-  decorationType: Decoration_Type.LINK,
   data1: LinkData.fromJSON({ url: 'www.wix.com', rel: 'nofollow', target: '_self' }),
   selection1: selection,
   data2: LinkData.fromJSON({ url: 'www.sport5.co.il', rel: 'sponsored', target: '_self' }),
@@ -503,7 +505,6 @@ const link = {
 
 const mention = {
   type: RICOS_MENTION_TYPE,
-  decorationType: Decoration_Type.MENTION,
   data1: { mention: MentionData.fromJSON({ name: 'aviv', slug: 'blabla' }), trigger: '@' },
   selection1: endOfSelection,
   data2: {
@@ -525,6 +526,26 @@ const mention = {
   },
 };
 
+const textColor = {
+  type: RICOS_TEXT_COLOR_TYPE,
+  data1: { color: 'color4' },
+  selection1: selection,
+  data2: { color: 'color2' },
+  selection2: selectionCollapsed,
+  expectedData1: 'color4',
+  expectedData2: 'color2',
+};
+
+const highlightColor = {
+  type: RICOS_TEXT_HIGHLIGHT_TYPE,
+  data1: { color: 'color3' },
+  selection1: selection,
+  data2: { color: 'color1' },
+  selection2: selectionCollapsed,
+  expectedData1: 'color3',
+  expectedData2: 'color1',
+};
+
 export const inlineStylesTestConfig = ['bold', 'italic', 'underline', 'spoiler'];
 
 export const pluginsTestConfig = {
@@ -538,4 +559,4 @@ export const pluginsTestConfig = {
   image,
 };
 
-export const decorationsTestConfig = { link, mention };
+export const decorationsTestConfig = { link, mention, textColor, highlightColor };