+++ packages/ui-components/web/src/index.ts
@@ -1,7 +1,9 @@
-export { FOOTER_BUTTON_ALIGNMENT, MODAL_CONTROLS_POSITION } from './consts';
+export { FOOTER_BUTTON_ALIGNMENT, MODAL_CONTROLS_POSITION, DECORATION_MODE } from './consts';
 
+// Modals
 export { default as UrlInputModal } from './Modals/UrlInputModal';
 export { default as SettingsMobileHeader } from './Modals/SettingsMobileHeader';
+export { default as RichContentModal } from './Modals/RichContentModal';
 
 export * from './Icons';
 
@@ -22,8 +24,13 @@ export { default as SliderWithInput } from './Components/SliderWithInput';
 export { default as Slider } from './Components/Slider';
 export { default as TextInput } from './Components/TextInput';
 export { Tab, Tabs } from './Components/Tabs';
-// export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 export { default as Loader } from './Components/Loader';
 export { default as MediaItemErrorMsg } from './Components/MediaItemErrorMsg';
-
+export { default as TextSearchInput } from './Components/TextSearchInput';
+export { default as Separator } from './Components/Separator';
+export { default as Dropdown } from './Components/Dropdown';
+export { default as InfoIcon } from './Components/InfoIcon';
+export { default as Checkbox } from './Components/Checkbox';
+export { default as RadioGroup } from './Components/RadioGroup';
+export { default as FocusManager } from './Components/FocusManager';
 // export { getSelectionStyles } from './Utils/inlineStyleUtils';